backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
httpx>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
openai>=1.3.0
pdfminer.six>=20221105
